fix(runtimestate): split inodeOf for non-Unix portability + M4 rename in test comment (Codex R7)
Codex R7 P3 (3215207340): "Guard the Unix-only inode type in the tagged
test". syscall.Stat_t is undefined on non-Unix Go targets (Windows,
Plan 9, etc.), so TestPersist_AtomicTempRename failed to compile under
GOOS=windows.

Fix: split inodeOf into platform-tagged files:
- runtimestate_test_unix.go (build tag: runtime_state_tdd_red && (linux ||
  darwin || freebsd || netbsd || openbsd || dragonfly || solaris)) —
  real implementation using syscall.Stat_t.Ino.
- runtimestate_test_other.go (build tag: runtime_state_tdd_red &&
  !linux && !darwin && ... ) — stub returning (0, false) so the
  caller t.Skip's the inode-distinguish assertion gracefully.

Also: corrected M4_JOINER_HINT reference to M4_SOURCE_SELECTION_HINT
in TestState_EBusSelfPresentDoesNotImplyAdmitted comment (companion to
the milestone rename in plan PR #28).

Verified: GOOS=darwin and GOOS=windows both compile cleanly. RED gate
still 17 RED + 1 SKIP under tdd_red tag.
